--- src/docs/asciidoc/account/post-200.adoc
@@ -1,37 +1,0 @@
-=== 가입 성공 (200)
-:basedir: {docdir}/../../../
-:snippets: {basedir}/build/generated-snippets
-:icons: font
-:source-highlighter: highlightjs
-:toc: left
-:toclevels: 4
-
-==== 설명
-
-신규 회원을 데이터베이스에 등록합니다.
-
-CAUTION: `Nullable` 응답 상세의 경우 값이 NULL이면 JSON에서 해당 필드가 제거됩니다.
-
-==== 요청
-
-===== 요청 상세
-
-include::{snippets}/account/post-200/request-fields.adoc[]
-
-===== 요청 예시
-
-include::{snippets}/account/post-200/http-request.adoc[]
-
-===== Curl
-
-include::{snippets}/account/post-200/curl-request.adoc[]
-
-==== 응답
-
-===== 응답 상세
-
-include::{snippets}/account/post-200/response-fields.adoc[]
-
-===== 응답 예시
-
-include::{snippets}/account/post-200/http-response.adoc[]
