+++ src/main/java/team/ark/sprout/adapter/in/web/ResponseData.java
@@ -1,0 +1,15 @@
+package team.ark.sprout.adapter.in.web;
+
+import lombok.Value;
+import org.springframework.http.HttpStatus;
+
+@Value(staticConstructor = "of")
+public class ResponseData<T> {
+    int code;
+    String message;
+    T data;
+
+    public static <T> ResponseData<T> create(HttpStatus status, T data) {
+        return new ResponseData<>(status.value(), status.getReasonPhrase(), data);
+    }
+}
