+++ src/main/java/team/ark/sprout/common/validate/SignUpFormValidator.java
@@ -1,0 +1,32 @@
+package team.ark.sprout.common.validate;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.validation.Errors;
+import org.springframework.validation.Validator;
+import team.ark.sprout.adapter.in.web.SignUpForm;
+import team.ark.sprout.adapter.out.persistence.AccountRepository;
+
+@Component
+@RequiredArgsConstructor
+public class SignUpFormValidator implements Validator {
+    private final AccountRepository repository;
+
+    @Override
+    public boolean supports(Class<?> clazz) {
+        return clazz.isAssignableFrom(SignUpForm.class);
+    }
+
+    @Override
+    public void validate(Object target, Errors errors) {
+        SignUpForm signUpForm = (SignUpForm) target;
+
+        if (repository.existsByEmail(signUpForm.getEmail())) {
+            errors.rejectValue("email", "invalid.email", new Object[]{signUpForm.getEmail()}, "이미 사용중인 이메일입니다");
+        }
+
+        if (repository.existsByNickName(signUpForm.getNickname())) {
+            errors.rejectValue("nickname", "invalid.nickname", new Object[]{signUpForm.getNickname()}, "이미 사용중인 닉네임입니다");
+        }
+    }
+}
