--- src/main/java/team/ark/sprout/config/DatabaseProperties.java
@@ -1,20 +1,0 @@
-package team.ark.sprout.config;
-
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-
-@Getter
-@Configuration
-@PropertySource("file:${user.home}/team-ark-backend/sprout-database-${spring.profiles.active}.properties")
-public class DatabaseProperties {
-    @Value("${spring.datasource.url}")
-    private String url;
-
-    @Value("${spring.datasource.username}")
-    private String username;
-
-    @Value("${spring.datasource.password}")
-    private String password;
-}
