--- src/main/java/team/ark/sprout/config/extension/ExtensionCreateAtEntity.java
@@ -1,40 +1,0 @@
-package team.ark.sprout.config.extension;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.io.Serializable;
-import java.time.LocalDateTime;
-import javax.persistence.Access;
-import javax.persistence.AccessType;
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import org.springframework.format.annotation.DateTimeFormat;
-
-/**
- * Entity의 최초 등록시점을 추가해야 할 경우 상속합니다.
- */
-@Getter
-@MappedSuperclass
-@Access(AccessType.FIELD)
-@EntityListeners(AuditingEntityListener.class)
-public abstract class ExtensionCreateAtEntity implements Serializable {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Long id;
-
-    @CreatedDate
-    @Column(updatable = false)
-    @DateTimeFormat(pattern = "yyyy-MM-dd")
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "Asia/Seoul")
-    protected LocalDateTime createdAt;
-
-    public boolean isNew() {
-        return this.id == null;
-    }
-}
