+++ src/main/java/team/ark/sprout/domain/account/AccountBasic.java
@@ -1,0 +1,10 @@
+package team.ark.sprout.domain.account;
+
+import lombok.Value;
+
+@Value(staticConstructor = "of")
+public class AccountBasic {
+    String email;
+    String nickname;
+    String password;
+}
