+++ src/main/java/team/ark/sprout/port/in/RegistAccount.java
@@ -1,0 +1,7 @@
+package team.ark.sprout.port.in;
+
+import team.ark.sprout.domain.account.Account;
+
+public interface RegistAccount {
+    Account regist(Account account);
+}
